+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## v1.5.0
+ - Show public DNS name next to the instances' zones.
  - Improve order of matched instances with similar scores.
 
 ## v1.4.0